+++ R/sanitize_gof_map.R
@@ -5,9 +5,32 @@ sanitize_gof_map <- function(gof_map) {
 
   # sanity checks
   checkmate::assert(
+    checkmate::check_character(gof_map, null.ok = TRUE, min.len = 1),
     checkmate::check_data_frame(gof_map, null.ok = TRUE, min.cols = 3),
     checkmate::check_list(gof_map, null.ok = TRUE),
     combine = "or")
+
+  # character vector
+  if (isTRUE(checkmate::check_character(gof_map))) {
+    known <- intersect(gof_map, modelsummary::gof_map$raw)
+    unknown <- setdiff(gof_map, modelsummary::gof_map$raw)
+    if (length(known) > 0) {
+      idx <- match(known, modelsummary::gof_map$raw)
+      known <- modelsummary::gof_map[idx, ]
+    } else {
+      known <- NULL
+    }
+    if (length(unknown) > 0) {
+      unknown <- data.frame(raw = unknown, clean = unknown, fmt = 0)
+    } else {
+      unknown <- NULL
+    }
+    gof_map <- bind_rows(known, unknown)
+    gof_map[["omit"]] <- FALSE
+  }
+
+
+  # data.frame
   if (inherits(gof_map, "data.frame")) {
     checkmate::assert_true(all(c("raw", "clean", "fmt") %in% colnames(gof_map)))
   } else if (class(gof_map)[1] == "list") {